Move cityrep* to views subdirectory with a file rename
