+++ client/views/view_cities.cpp
@@ -1,11 +1,16 @@
 /*
- Copyright (c) 1996-2022 Freeciv21 and Freeciv contributors. This file is
+ Copyright (c) 1996-2023 Freeciv21 and Freeciv contributors. This file is
  part of Freeciv21. Freeciv21 is free software: you can redistribute it
  and/or modify it under the terms of the GNU  General Public License  as
  published by the Free Software Foundation, either version 3 of the
  License,  or (at your option) any later version. You should have received
  a copy of the GNU General Public License along with Freeciv21. If not,
  see https://www.gnu.org/licenses/.
+ */
+
+/*
+ * This file contains functions to generate the table based GUI for
+ * the city view (formally known as the city report).
  */
 
 // Qt
@@ -23,7 +28,6 @@
 #include "client_main.h"
 #include "governor.h"
 // gui-qt
-#include "cityrep.h"
 #include "fc_client.h"
 #include "hudwidget.h"
 #include "icons.h"
@@ -31,6 +35,7 @@
 #include "page_game.h"
 #include "qtg_cxxside.h"
 #include "top_bar.h"
+#include "views/view_cities.h"
 
 /**
    Overriden compare for sorting items
