+++ client/views/view_cities.h
@@ -1,5 +1,5 @@
 /**************************************************************************
- Copyright (c) 1996-2020 Freeciv21 and Freeciv contributors. This file is
+ Copyright (c) 1996-2023 Freeciv21 and Freeciv contributors. This file is
  part of Freeciv21. Freeciv21 is free software: you can redistribute it
  and/or modify it under the terms of the GNU  General Public License  as
  published by the Free Software Foundation, either version 3 of the
@@ -17,8 +17,8 @@
 #include <QTreeView>
 #include <QWidget>
 // client
-#include "cityrepdata.h"
 #include "climisc.h"
+#include "views/view_cities_data.h"
 
 #define CMA_NONE (10000)
