+++ client/views/view_cities_data.h
